Pattern matching is starting to work now.
Other than fixing some bugs and enabling stuff like automatic function
signing (when it's indicated, such as in a -> (b -> c) where the return
value has to be signed. Constructors work a bit differently and pattern
matching works at a very early stage. The only issue now is that Nothing
is not matched as Maybe a, so in the next commit that will probably be
fixed.